+++ docs/changes.md
@@ -3,6 +3,8 @@
   * 增加了object\_get\_type等函数。
   * 修改widget\_use\_style的问题。
   * 修改emitter\_remove\_item的BUG。
+  * 增加object\_default\_clone。
+  * 修改widget\_clone没有拷贝自定义属性的问题。
 
 * 2019/03/12
   * 窗口管理器增加window\_manager\_back/window\_manager\_back\_to\_home